+++ src/test/converter2/issues/gh3070.ts
@@ -1,0 +1,3 @@
+export class NumberArray extends Array<number> {
+    foo() {}
+}