+++ src/test/issues.c2.test.ts
@@ -2304,6 +2304,12 @@ describe("Issue Tests", () => {
         equal(Comment.combineDisplayParts(project.children[1].signatures?.[0].comment?.summary), "B");
     });
 
+    it("#3070 does not warn due to bad comment in declaration file", () => {
+        app.options.setValue("excludeExternals", false);
+        app.options.setValue("suppressCommentWarningsInDeclarationFiles", true);
+        convert();
+    });
+
     it("#3071 correctly handles substitution types", () => {
         const project = convert();
         const m = querySig(project, "m");