+++ CHANGELOG.md
@@ -9,6 +9,8 @@ title: Changelog
 - Improved handling of comments for type aliases which have been declaration merged with functions, #3064.
 - Fixed anchor link generation to members named `$`, #3065.
 - Corrected typing of the `plugin` option to permit functions, #3066.
+- Warnings about unused `@param` tags will now be properly suppressed when they come from declaration files and
+  the suppressCommentWarningsInDeclarationFiles option is enabled, #3070.
 - Fixed conversion of types referencing type parameters on functions, #3071.
 
 ### Thanks!